Added a ReceptionAndAssignment step to restriction cases
