+++ apps/judicial-system/web/messages/RestrictionCases/Court/receptionAndAssignment.ts
@@ -1,0 +1,10 @@
+import { defineMessage } from 'react-intl'
+
+export const rcReceptionAndAssignment = {
+  title: defineMessage({
+    id: 'judicial.system.restriction_cases:reception_and_assignment.title',
+    defaultMessage: 'Móttaka og úthlutun',
+    description:
+      'Notaður sem titill á Móttaka og úthlutun skjá í gæsluvarðhalds- og farbannsmálum.',
+  }),
+}
