+++ apps/judicial-system/web/pages/domur/krafa/[id].ts
