+++ apps/judicial-system/web/pages/domur/mottaka/[id].ts
@@ -1,0 +1,3 @@
+import ReceptionAndAssignment from '@island.is/judicial-system-web/src/routes/Court/RestrictionRequest/ReceptionAndAssignment/ReceptionAndAssignment'
+
+export default ReceptionAndAssignment
