+++ apps/judicial-system/web/pages/domur/urskurdur/[id].ts
