+++ apps/judicial-system/web/pages/index.ts
