+++ apps/judicial-system/web/pages/krafa/fyrirtaka/[id].ts
