+++ apps/judicial-system/web/pages/krafa/sakborningur/[id].ts
