+++ apps/judicial-system/web/pages/krafa/stadfesta/[id].ts
