+++ apps/judicial-system/web/pages/krafa/yfirlit/[id].ts
