+++ apps/judicial-system/web/pages/krofur.ts
