+++ apps/judicial-system/web/src/routes/Court/RestrictionRequest/Overview/Overview.tsx
@@ -7,16 +7,9 @@ import {
   FormContentContainer,
 } from '@island.is/judicial-system-web/src/components'
 import {
-  CaseState,
-  CaseTransition,
-  NotificationType,
-} from '@island.is/judicial-system/types'
-import type { Case } from '@island.is/judicial-system/types'
-import {
   JudgeSubsections,
   Sections,
 } from '@island.is/judicial-system-web/src/types'
-import { useCase } from '@island.is/judicial-system-web/src/utils/hooks'
 import { isOverviewStepValidRC } from '@island.is/judicial-system-web/src/utils/validate'
 import { FormContext } from '@island.is/judicial-system-web/src/components/FormProvider/FormProvider'
 import * as Constants from '@island.is/judicial-system-web/src/utils/constants'
@@ -32,54 +25,14 @@ export const JudgeOverview: React.FC = () => {
     caseNotFound,
     isCaseUpToDate,
   } = useContext(FormContext)
-  const [courtCaseNumberEM, setCourtCaseNumberEM] = useState('')
   const [isDraftingConclusion, setIsDraftingConclusion] = useState<boolean>()
-  const [createCourtCaseSuccess, setCreateCourtCaseSuccess] = useState<boolean>(
-    false,
-  )
 
   const router = useRouter()
   const id = router.query.id
 
-  const {
-    createCourtCase,
-    isCreatingCourtCase,
-    transitionCase,
-    isTransitioningCase,
-    sendNotification,
-  } = useCase()
-
   useEffect(() => {
     document.title = 'Yfirlit kröfu - Réttarvörslugátt'
   }, [])
-
-  const receiveCase = async (workingCase: Case, courtCaseNumber: string) => {
-    if (workingCase.state === CaseState.SUBMITTED && !isTransitioningCase) {
-      // Transition case from SUBMITTED to RECEIVED when courtCaseNumber is set
-      const received = await transitionCase(
-        { ...workingCase, courtCaseNumber },
-        CaseTransition.RECEIVE,
-        setWorkingCase,
-      )
-
-      if (received) {
-        sendNotification(workingCase.id, NotificationType.RECEIVED_BY_COURT)
-      }
-    }
-  }
-
-  const handleCreateCourtCase = async (workingCase: Case) => {
-    const courtCaseNumber = await createCourtCase(
-      workingCase,
-      setWorkingCase,
-      setCourtCaseNumberEM,
-    )
-
-    if (courtCaseNumber !== '') {
-      setCreateCourtCaseSuccess(true)
-      receiveCase(workingCase, courtCaseNumber)
-    }
-  }
 
   return (
     <PageLayout
@@ -94,14 +47,7 @@ export const JudgeOverview: React.FC = () => {
       <OverviewForm
         workingCase={workingCase}
         setWorkingCase={setWorkingCase}
-        handleCreateCourtCase={handleCreateCourtCase}
-        createCourtCaseSuccess={createCourtCaseSuccess}
-        setCreateCourtCaseSuccess={setCreateCourtCaseSuccess}
-        courtCaseNumberEM={courtCaseNumberEM}
-        setCourtCaseNumberEM={setCourtCaseNumberEM}
         setIsDraftingConclusion={setIsDraftingConclusion}
-        isCreatingCourtCase={isCreatingCourtCase}
-        receiveCase={receiveCase}
       />
       <FormContentContainer isFooter>
         <FormFooter
